Labs/discovery-lab/scripts/requirements.txt
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
docker==7.1.0
future==1.0.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
netmiko==4.5.0
ntc_templates==7.7.0
packaging==24.2
paramiko==3.5.1
pycparser==2.22
Pygments==2.19.1
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==1.0.1
PyYAML==6.0.2
pyzabbix==1.3.1
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scp==0.15.0
six==1.17.0
textfsm==1.1.3
typing_extensions==4.12.2
urllib3==2.3.0
